src/aimodels/custom_vision_object_detection_offline/src/requirements.txt
numpy~=1.17.5 
tensorflow~=2.0.2 
flask~=1.1.2 
pillow~=7.2.0
mscviplib==2.200731.16 
MarkupSafe~=2.0.1